+++ lib/iex/resources/symbol.rb
@@ -1,0 +1,10 @@
+module IEX
+  module Resources
+    class Symbol < Resource
+      property 'exchange'
+      property 'iex_id', from: 'iexId'
+      property 'region'
+      property 'symbol'
+    end
+  end
+end