+++ spec/iex/endpoints/ref_data_spec.rb
@@ -3,14 +3,30 @@ require 'spec_helper'
 describe IEX::Api::Client do
   include_context 'client'
 
-  describe 'ISIN Mapping', vcr: { cassette_name: 'ref-data/isin' } do
-    subject do
-      client.post('ref-data/isin', isin: ['US0378331005'], token: client.secret_token)
+  describe '#ref_data_isin', vcr: { cassette_name: 'ref-data/isin' } do
+    context 'without options' do
+      subject { client.ref_data_isin(['US0378331005']) }
+
+      it 'converts ISIN to IEX Cloud symbols' do
+        expect(subject.count).to eq(2)
+        expect(subject.first).to eq('exchange' => 'NAS', 'iex_id' => 'IEX_4D48333344362D52', 'region' => 'US', 'symbol' => 'AAPL')
+        expect(subject.last).to eq('exchange' => 'ETR', 'iex_id' => 'IEX_464D46474C312D52', 'region' => 'DE', 'symbol' => 'APC-GY')
+      end
     end
-    it 'retrieves a ticker by ISIN' do
-      expect(subject.count).to eq(2)
-      expect(subject.first).to eq('exchange' => 'NAS', 'iexId' => 'IEX_4D48333344362D52', 'region' => 'US', 'symbol' => 'AAPL')
-      expect(subject.last).to eq('exchange' => 'ETR', 'iexId' => 'IEX_464D46474C312D52', 'region' => 'DE', 'symbol' => 'APC-GY')
+
+    context 'with mapped option', vcr: { cassette_name: 'ref-data/isin_mapped' } do
+      subject { client.ref_data_isin(%w[US0378331005 US5949181045], mapped: true) }
+
+      it 'converts ISINs to IEX Cloud symbols mapped by ISIN' do
+        expect(subject.keys).to contain_exactly('US0378331005', 'US5949181045')
+
+        expect(subject['US0378331005'].first).to eq('exchange' => 'NAS', 'iex_id' => 'IEX_4D48333344362D52', 'region' => 'US', 'symbol' => 'AAPL')
+        expect(subject['US0378331005'].last).to eq('exchange' => 'ETR', 'iex_id' => 'IEX_464D46474C312D52', 'region' => 'DE', 'symbol' => 'APC-GY')
+
+        expect(subject['US5949181045'][0]).to eq('exchange' => 'NAS', 'iex_id' => 'IEX_5038523343322D52', 'region' => 'US', 'symbol' => 'MSFT')
+        expect(subject['US5949181045'][1]).to eq('exchange' => 'ETR', 'iex_id' => 'IEX_4C42583859482D52', 'region' => 'DE', 'symbol' => 'MSF-GY')
+        expect(subject['US5949181045'][2]).to eq('exchange' => 'BRU', 'iex_id' => 'IEX_5833345950432D52', 'region' => 'BE', 'symbol' => 'MSF-BB')
+      end
     end
   end
 end